+++ lib/components/Link/Link.playroom.tsx
@@ -1,10 +1,13 @@
-import React from 'react';
+import React, { forwardRef } from 'react';
 import { Link as CustomLink, LinkProps } from './Link';
 
-export const Link = ({ href, onClick, ...restProps }: LinkProps) => (
-  <CustomLink
-    href={href ?? ''}
-    onClick={onClick ? onClick : (event) => event?.preventDefault()}
-    {...restProps}
-  />
+export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
+  ({ href, onClick, ...restProps }, ref) => (
+    <CustomLink
+      ref={ref}
+      href={href ?? ''}
+      onClick={onClick ? onClick : (event) => event?.preventDefault()}
+      {...restProps}
+    />
+  ),
 );