+++ lib/components/MenuItem/MenuItem.tsx
@@ -1,29 +1,23 @@
+import assert from 'assert';
 import React, {
   KeyboardEvent,
   useContext,
   useRef,
   useEffect,
   ReactNode,
-  MouseEvent,
 } from 'react';
 import { useStyles } from 'sku/react-treat';
-import assert from 'assert';
-import { Box } from '../Box/Box';
+import { Box, Text, Link } from '../';
+import { LinkProps } from '../Link/Link';
 import { useTouchableSpace } from '../../hooks/typography';
 import { normalizeKey } from '../private/normalizeKey';
 import { MenuContext } from '../MenuRenderer/MenuRenderer';
 import { actionTypes, Action } from '../MenuRenderer/MenuRenderer.actions';
-import * as styleRefs from './MenuItem.treat';
-import { Text } from '../Text/Text';
 import buildDataAttributes, {
   DataAttributeMap,
 } from '../private/buildDataAttributes';
-
-interface MenuItemProps {
-  children: ReactNode;
-  onClick?: () => void;
-  data?: DataAttributeMap;
-}
+import * as styleRefs from './MenuItem.treat';
+import { useBoxStyles } from '../Box/useBoxStyles';
 
 const {
   MENU_ITEM_UP,
@@ -36,35 +30,61 @@ const {
   MENU_ITEM_HOVER,
 } = actionTypes;
 
-export const MenuItem = ({ children, onClick, data }: MenuItemProps) => {
+const menuItemChildrenSize = 'standard';
+
+interface UseMenuItemProps {
+  onClick?: () => void;
+  data?: DataAttributeMap;
+  displayName?: string;
+}
+function useMenuItem<MenuItemElement extends HTMLElement>({
+  displayName = 'MenuItem',
+  onClick,
+  data,
+}: UseMenuItemProps) {
   const styles = useStyles(styleRefs);
   const menuContext = useContext(MenuContext);
 
   assert(
     menuContext !== null,
-    'A MenuItem must be rendered as an immediate child of a Menu. See the documentation for correct usage: https://seek-oss.github.io/braid-design-system/components/MenuRenderer',
+    `${displayName} must be rendered as an immediate child of a menu. See the documentation for correct usage: https://seek-oss.github.io/braid-design-system/components/MenuItem`,
   );
 
   if (menuContext === null) {
-    throw new Error('MenuItem rendered outside menu context');
+    throw new Error(`${displayName} element rendered outside menu context`);
   }
 
   const { isHighlighted, index, dispatch, focusTrigger } = menuContext;
-  const menuItemRef = useRef<HTMLButtonElement>(null);
+  const menuItemRef = useRef<MenuItemElement>(null);
 
   useEffect(() => {
-    if (menuItemRef.current && isHighlighted) {
-      menuItemRef.current.focus();
+    if (isHighlighted) {
+      menuItemRef.current?.focus();
     }
   }, [isHighlighted]);
 
-  const clickHandler = () => {
-    if (typeof onClick === 'function') {
-      onClick();
+  const onKeyDown = (event: KeyboardEvent<MenuItemElement>) => {
+    const targetKey = normalizeKey(event);
+
+    if (targetKey === 'Tab') {
+      dispatch({ type: MENU_ITEM_TAB });
+    }
+
+    const isArrowPress = targetKey.indexOf('Arrow') === 0;
+    const isActionKeyPress = targetKey === 'Enter' || targetKey === ' ';
+
+    // This prevents spacebar from scrolling the document,
+    // but also prevents the click event from firing so we
+    // can programmatically trigger a click event in the
+    // 'onKeyUp' handler. This is to normalise behaviour
+    // between buttons and links, e.g. to make spacebar
+    // activate links, which is not standard behaviour.
+    if (isArrowPress || isActionKeyPress) {
+      event.preventDefault();
     }
   };
 
-  const onKeyUp = (event: KeyboardEvent<HTMLButtonElement>) => {
+  const onKeyUp = (event: KeyboardEvent<MenuItemElement>) => {
     const targetKey = normalizeKey(event);
     const closeActionKeys = ['Enter', ' ', 'Escape'];
 
@@ -80,8 +100,10 @@ export const MenuItem = ({ children, onClick, data }: MenuItemProps) => {
       dispatch(action[targetKey]);
     }
 
+    // Because we call 'preventDefault()' on enter/spacebar in
+    // the 'onKeyDown' handler, we manually trigger a click here.
     if (targetKey === 'Enter' || targetKey === ' ') {
-      clickHandler();
+      menuItemRef.current?.click();
     }
 
     if (closeActionKeys.indexOf(targetKey) > -1) {
@@ -89,57 +111,81 @@ export const MenuItem = ({ children, onClick, data }: MenuItemProps) => {
     }
   };
 
-  const onKeyDown = (event: KeyboardEvent<HTMLButtonElement>) => {
-    const targetKey = normalizeKey(event);
+  return {
+    role: 'menuitem',
+    tabIndex: -1,
+    ref: menuItemRef,
+    onKeyUp,
+    onKeyDown,
+    onMouseEnter: () => dispatch({ type: MENU_ITEM_HOVER, value: index }),
+    onClick: () => {
+      dispatch({ type: MENU_ITEM_CLICK });
 
-    if (targetKey === 'Tab') {
-      dispatch({ type: MENU_ITEM_TAB });
-    }
+      if (typeof onClick === 'function') {
+        onClick();
+      }
+    },
+    className: [
+      styles.menuItem,
+      useTouchableSpace(menuItemChildrenSize),
+      useBoxStyles({
+        component: null,
+        display: 'block',
+        width: 'full',
+        paddingX: 'small',
+        background: isHighlighted ? 'selection' : undefined,
+        cursor: 'pointer',
+        textAlign: 'left',
+        outline: 'none',
+      }),
+    ],
+    ...buildDataAttributes(data),
+  } as const;
+}
 
-    // Prevent arrow keys scrolling the document while navigating the menu
-    const isArrowPress = targetKey.indexOf('Arrow') === 0;
-    // Prevent enter or space press from triggering the click handler
-    const isActionKeyPress = targetKey === 'Enter' || targetKey === ' ';
+interface MenuItemChildrenProps {
+  children: ReactNode;
+}
+const MenuItemChildren = ({ children }: MenuItemChildrenProps) => (
+  <Text size={menuItemChildrenSize} baseline={false}>
+    {children}
+  </Text>
+);
 
-    if (isArrowPress || isActionKeyPress) {
-      event.preventDefault();
-    }
-  };
+interface MenuItemProps extends Omit<UseMenuItemProps, 'displayName'> {
+  children: ReactNode;
+}
+export const MenuItem = ({ children, onClick, data }: MenuItemProps) => (
+  <Box
+    {...useMenuItem<HTMLButtonElement>({ onClick, data })}
+    component="button"
+    type="button"
+  >
+    <MenuItemChildren>{children}</MenuItemChildren>
+  </Box>
+);
 
-  const menuItemTextSize = 'standard';
-
-  return (
-    <Box
-      component="button"
-      role="menuitem"
-      tabIndex={-1}
-      ref={menuItemRef}
-      onKeyUp={onKeyUp}
-      onKeyDown={onKeyDown}
-      onMouseEnter={() => dispatch({ type: MENU_ITEM_HOVER, value: index })}
-      onClick={(event: MouseEvent) => {
-        event.stopPropagation();
-        event.preventDefault();
-        dispatch({ type: MENU_ITEM_CLICK });
-        clickHandler();
-      }}
-      display="block"
-      width="full"
-      paddingX="small"
-      background={isHighlighted ? 'selection' : undefined}
-      cursor="pointer"
-      textAlign="left"
-      outline="none"
-      className={[useTouchableSpace(menuItemTextSize), styles.menuItem]}
-      {...buildDataAttributes(data)}
-    >
-      {/*
-        Rendering Text component to provide rendering context
-        for both icons and text labels
-      */}
-      <Text size={menuItemTextSize} baseline={false}>
-        {children}
-      </Text>
-    </Box>
-  );
-};
+interface MenuItemLinkProps
+  extends MenuItemProps,
+    Pick<LinkProps, 'href' | 'target' | 'rel'> {}
+export const MenuItemLink = ({
+  href,
+  target,
+  rel,
+  onClick,
+  data,
+  children,
+}: MenuItemLinkProps) => (
+  <Link
+    {...useMenuItem<HTMLAnchorElement>({
+      displayName: 'MenuItemLink',
+      onClick,
+      data,
+    })}
+    href={href}
+    target={target}
+    rel={rel}
+  >
+    <MenuItemChildren>{children}</MenuItemChildren>
+  </Link>
+);