+++ lib/components/TextLink/TextLink.test.tsx
@@ -2,7 +2,13 @@ import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import { render } from '@testing-library/react';
 import { htmlToText } from '../../utils/htmlToText';
-import { BraidTestProvider, LinkComponent, Text, TextLink } from '..';
+import {
+  BraidTestProvider,
+  makeLinkComponent,
+  LinkComponent,
+  Text,
+  TextLink,
+} from '..';
 
 describe('TextLink', () => {
   it('should render a native link by default', () => {
@@ -24,6 +30,29 @@ describe('TextLink', () => {
   });
 
   it('should render a custom link component if provided', () => {
+    const CustomLink = makeLinkComponent((props, ref) => (
+      <a ref={ref} {...props} data-custom-link-component="true" />
+    ));
+
+    const { getByRole } = render(
+      <BraidTestProvider linkComponent={CustomLink}>
+        <Text>
+          <TextLink href="/foo/bar" data-attribute="true">
+            Link content
+          </TextLink>
+        </Text>
+      </BraidTestProvider>,
+    );
+
+    const link = getByRole('link');
+    expect(link.nodeName).toEqual('A');
+    expect(link.getAttribute('href')).toEqual('/foo/bar');
+    expect(htmlToText(link.innerHTML)).toEqual('Link content');
+    expect(link.getAttribute('data-attribute')).toEqual('true');
+    expect(link.getAttribute('data-custom-link-component')).toEqual('true');
+  });
+
+  it('should render a legacy link component if provided', () => {
     const CustomLink: LinkComponent = (props) => (
       <a {...props} data-custom-link-component="true" />
     );
@@ -47,9 +76,9 @@ describe('TextLink', () => {
   });
 
   it("should inherit custom link components from the root provider if the nearest provider doesn't have one", () => {
-    const CustomLink: LinkComponent = (props) => (
-      <a {...props} data-custom-link-component="true" />
-    );
+    const CustomLink = makeLinkComponent((props, ref) => (
+      <a ref={ref} {...props} data-custom-link-component="true" />
+    ));
 
     const { getByRole } = render(
       <BraidTestProvider linkComponent={CustomLink}>