+++ lib/components/private/zIndex.ts
@@ -1,6 +1,0 @@
-export const status = 600;
-export const headerOverlay = 500;
-export const header = 400;
-export const pageOverlay = 300;
-export const backdrop = 200;
-export const inlineOverlay = 100;