+++ lib/components/useToast/Toast.treat.ts
@@ -1,11 +1,6 @@
 import { style } from 'sku/treat';
-import { status } from '../private/zIndex';
 
 export const toast = style({
   pointerEvents: 'all',
   maxWidth: 400,
-});
-
-export const toaster = style({
-  zIndex: status,
 });