+++ lib/themes/seekBusiness/tokens.ts
@@ -5,5 +5,5 @@ export default makeTokens({
   displayName: 'SEEK Business',
   brand: '#009fd4',
   brandAccent: '#0d3880',
-  formAccent: '#009fd4',
+  formAccent: '#2765cf',
 });